Update Gallery Block
Updated Gallery block to use first part of breadcrumb due to all link
being set to it's own url. This now sets it to the first part of the
breadcrumbs. You can expect this to cause issues when the gallery
component is used when inside of a nested page itself.

For example: /page/gallery-block/another-gallery-block.